+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/controller/model/VariantDefinition.kt
@@ -10,4 +10,5 @@ data class VariantDefinition(
   val printable: Boolean? = true,
   val summaries: List<ReportSummary>? = emptyList(),
   val interactive: Boolean? = null,
+  val childVariants: List<ChildVariantDefinition>? = null,
 )