+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/Dataset.kt
@@ -7,4 +7,6 @@ data class Dataset(
   val query: String,
   val schema: Schema,
   val parameters: List<Parameter>? = null,
-)
+) : Identified() {
+  override fun getIdentifier() = this.id
+}