+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/SchemaField.kt
@@ -5,4 +5,6 @@ data class SchemaField(
   val type: ParameterType,
   val display: String,
   val filter: FilterDefinition?,
-)
+) : Identified() {
+  override fun getIdentifier() = this.name
+}