+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/SingleReportProductDefinition.kt
@@ -13,4 +13,5 @@ data class SingleReportProductDefinition(
   val report: Report,
   override val policy: List<Policy>,
   val allDatasets: List<Dataset>,
+  val allReports: List<Report>,
 ) : WithPolicy