+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/SummaryField.kt
@@ -4,4 +4,6 @@ data class SummaryField(
   val name: String,
   val header: Boolean? = false,
   val mergeRows: Boolean? = false,
-)
+) : Identified() {
+  override fun getIdentifier() = this.name
+}