+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/Template.kt
@@ -17,4 +17,7 @@ enum class Template {
 
   @SerializedName("summary-section")
   SectionedSummary,
+
+  @SerializedName("parent-child")
+  ParentChild,
 }