+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/DynamoDbProductDefinitionRepositoryTest.kt
@@ -28,6 +28,7 @@ class DynamoDbProductDefinitionRepositoryTest {
     dynamoDbClient = dynamoDbClient,
     gson = DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
     properties = properties,
+    identifiedHelper = IdentifiedHelper(),
   )
 
   @BeforeEach