+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 Below you can find the changes included in each release.
 
+# 7.5.0
+Add Parent-Child template: https://github.com/ministryofjustice/hmpps-digital-prison-reporting-data-product-definitions-schema?tab=readme-ov-file#parent-child-template
+
 # 7.4.0
-The dashboard result endpoint now returns a ResponseEntity<List<List<Map<String, Any?>>>> instead a ResponseEntity<List<Map<String, Any?>>> which was previously returned.  
+The dashboard result endpoint now returns a ResponseEntity<List<List<Map<String, Any?>>>> instead a ResponseEntity<List<Map<String, Any?>>> which was previously returned.
 
 # 7.3.21
 The status endpoints check if a table is missing only if the status is 'FINISHED' and there is a table ID provided.